# 🚀 Enhanced Tooling, New Features, Free LLM Support, Critical Bugfixes  (v2.5.1)