--- a/.env.example
+++ b/.env.example
@@ -36,6 +36,9 @@ OPENROUTER_API_KEY=YOUR_OPENROUTER_API_KEY_HERE
 # SSE Server Port (default: 3000)
 # SSE_PORT=3000
 
+# HTTP Agent Port (default: 3011)
+# HTTP_AGENT_PORT=3011
+
 # =============================================================================
 # DIRECTORY CONFIGURATION
 # =============================================================================