--- a/.gitignore
+++ b/.gitignore
@@ -40,6 +40,7 @@ test/
 */__tests__/
 */tests/
 */__integration__/
+tests/
 
 # IDE - VSCode
 .vscode/*