--- a/src/services/http-agent-api/index.ts
+++ b/src/services/http-agent-api/index.ts
@@ -437,7 +437,7 @@ class HTTPAgentAPIServer {
     }
   }
 
-  async start(port: number): Promise<void> {
+  async start(port: number = 3001): Promise<void> {
     try {
       // Validate port parameter (should be pre-allocated by Transport Manager)
       if (!port || port <= 0 || port > 65535) {