--- a/src/tools/vibe-task-manager/types/index.ts
+++ b/src/tools/vibe-task-manager/types/index.ts
@@ -3,3 +3,4 @@ export * from './task.js';
 export * from './dependency.js';
 export * from './agent.js';
 export * from './nl.js';
+export * from './artifact-types.js';