… README.md
… VIBE_CODER_MCP_SYSTEM_INSTRUCTIONS.md
… debug/DEBUG_README.md
… debug/VERSION_2_5_1_TESTING_SUMMARY.md
… debug/my-llm-config.json
… docs/ARCHITECTURE.md
… llm_config.json
… mcp-config.json
… package.json
… setup.sh
… src/index.ts
… src/logger.js
… src/logger.ts
… src/services/transport-manager/__tests__/dynamic-ports.test.ts
… src/services/transport-manager/index.ts
… src/tools/__tests__/dynamic-port-integration.test.ts
… src/tools/agent-registry/index.ts
… src/tools/agent-response/index.ts
… src/tools/agent-tasks/index.ts
… src/tools/code-map-generator/cache/__tests__/memoryManager.recursion.test.ts
… src/tools/code-map-generator/cache/fileCache.ts
… src/tools/code-map-generator/cache/memoryManager.ts
… src/tools/code-map-generator/configValidator.ts
… src/tools/code-map-generator/fsUtils.ts
… src/tools/code-map-generator/graphBuilder.ts
… src/tools/context-curator/__tests__/unit/types/context-curator.test.ts
… src/tools/context-curator/index.ts
… src/tools/context-curator/services/context-curator-service.ts
… src/tools/context-curator/types/context-curator.ts
… src/tools/job-result-retriever/index.ts