@evx2 2025-06-28
Updated to 2.5.1 which solved the infinite loop causing map-codebase to hang at 60%. Now, map-codebase works identically to old version when LLM is not available, but still retains all new features when LLM is available.